Added abstraction for key management
Instead of having to initialize different class for different API
levels, having an interface and abstract class that the key manager
class can extend makes it easier to use the KeyManager class without
having to modify how it is called in the classes that uses it.

Without the interface and abstract class, we would have to instantiate
different key manager class depending on the API level in each of the
class that needs to use the functions of a key manager.